backend/requirements.txt
flask
python-dotenv
ibm-watson
google-generativeai
pandas
pytest
